jeonse_fraud/requirements.txt
openai>=1.0.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0
loguru>=0.7.2
# For CLI and logging
argparse
# For Jupyter/Notebook (optional)
jupyter 